chore: Update index.html and vercel.json for security enhancements
This commit updates the index.html file in the Frontend/public directory to comment out the favicon link. This change is made to improve security by preventing the loading of the favicon.ico file from the public URL.

Additionally, the vercel.json file in the backend directory is modified to add a Content-Security-Policy header with the value 'default-src 'none''. This enhances security by specifying that no resources should be loaded from any source by default.

These changes aim to improve the security of the application.